utils: move to seperate file
